+++ frontend/src/app/features/main-page/components/add-categories-form/add-categories-form.component.ts
@@ -3,6 +3,10 @@ import { HttpClient } from '@angular/common/http';
 import { Component, Input, OnInit, ChangeDetectionStrategy } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 
+import { EMPTY, take } from 'rxjs';
+import { catchError } from 'rxjs/operators';
+
+import { ViewState } from './../../../../enums/view-state.enum';
 import { CategoryType } from './../../../../enums/category-type.enum';
 import { Category } from './../../../../interfaces/category.interface';
 
@@ -19,6 +23,8 @@ export class AddCategoriesFormComponent implements OnInit {
     public headingText: string = 'Add categories';
     public form: FormGroup = new FormGroup({});
     public CategoryType: typeof CategoryType = CategoryType;
+    public ViewState: typeof ViewState = ViewState;
+    public viewState: ViewState;
     public shouldShowCategories: boolean;
     public headingId: string = 'heading';
 
@@ -30,15 +36,28 @@ export class AddCategoriesFormComponent implements OnInit {
 
     public iterateByOriginalOrder = (): number => 0;
 
-    // TODO add states when backend is ready
+    // TODO: Add snackbars
     public saveCategories(): void {
         if (this.form.invalid) {
             return;
         }
 
+        this.viewState = ViewState.LOADING;
+
         const requestBody: Omit<Category, 'id'>[] = this.prepareCategoriesDataForRequest();
 
-        this.httpClient.post<Category[]>('categories', requestBody);
+        this.httpClient.post<Category[]>('categories', requestBody)
+            .pipe(
+                take(1),
+                catchError(() => {
+                    this.viewState = ViewState.ERROR;
+
+                    return EMPTY;
+                })
+            )
+            .subscribe(() => {
+                this.viewState = ViewState.SUCCESS;
+            });
     }
 
     public handleCategoriesVisibility(): void {